Add required libraries to test GPU implementation
Previously, the only way to test the GPU implementation was to install
it locally using an available OpenCL implementation installed on the
host.  Now, the Fedora and Ubuntu docker images support running the
tests via the Portable Compute Language (POCL) implementation of OpenCL.
This enables testing the implementation on platforms that only have
CPUs.